Add idea: MAKE IT BLUE AND YELLOW LIKE UKRAINE FLAG
we are harvard mit students 

we are developing scalable course that teaches how to do rapid prototpying usign latest ai tools 

for this project by decemember we are focussing on creating a hackathon for kids at school in their final year of school 
to guide them empower them to possible applications of ai and build them closely to rapidly moving technology 

we are planning to make a hakacathon in ma boston and later we are planning ot go to ukraine and do the same course there. 

so ma is our testing ground. 

what we are hoping to build in the future 
is a school which teaches and brings ai to any country or race or shape and shows / unlocks their creative potential through ai - to make and change the world a better place. 

we are asking for teaching assistants and money and connections to make this project happen between ma and ukraine. 

we want peopel in leading tech industries - from api credits to platforms with free stuff 
mentors from leading cutting edge places - openai hugging face google and cool places 
people who want to participate and shape the future generation 
role models 
indepdenent hacker who have built stuff... 

indie hackers? who really built somethign and who broke the rules and build projects alone 
who built themselves from scaratch 
no matter where you come from - you can make a difference 
independent thinkers. 

please donate and support this campaign here